requirements.txt
telethon==1.34.0
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
aiosqlite==0.19.0
cryptg==0.4.0
python-multipart==0.0.6
